flags.py: removed db usage and using normalStore for storage
