+++ twinTrim/flags.py
@@ -2,8 +2,8 @@ import os
 import click
 import time
 from collections import defaultdict
-from twinTrim.utils import find_duplicates,handle_and_remove
-from twinTrim.flagController import handleAllFlag
+from twinTrim.utils import handle_and_remove
+from twinTrim.flagController import handleAllFlag, find_duplicates
 from twinTrim.db import drop_all_tables, recreate_database
 
 @click.command()
